+++ yivi_core/lib/src/screens/pin/unsecure_pin_warning_text_button.dart
@@ -20,7 +20,7 @@ class _UnsecurePinWarningTextButton extends StatelessWidget {
       Navigator.of(context).push(
         PageRouteBuilder(
           opaque: false,
-          pageBuilder: (BuildContext context, _, __) =>
+          pageBuilder: (BuildContext context, _, _) =>
               _UnsecurePinFullScreen(state: state),
         ),
       );