+++ yivi_core/lib/src/screens/settings/settings_screen.dart
@@ -237,6 +237,7 @@ class _SettingsScreenState extends State<SettingsScreen> {
                   ),
                 ],
               ),
+              const SizedBox(height: 200),
             ],
           ),
         ),